src/broker/src/broker.rs
use rafka_core::{Message, Result, ThreadPool, Config};
use dashmap::DashMap;
use std::sync::Arc;
use tokio::sync::mpsc;

pub struct StatelessBroker {
    active_producers: DashMap<String, mpsc::Sender<Message>>,
    active_consumers: DashMap<String, Vec<mpsc::Sender<Message>>>,
    thread_pool: Arc<ThreadPool>,
    config: Config,
}

impl StatelessBroker {
    pub fn new(config: Config) -> Self {
        StatelessBroker {
            active_producers: DashMap::new(),
            active_consumers: DashMap::new(),
            thread_pool: Arc::new(ThreadPool::new(32)),
            config,
        }
    }

    pub async fn route_message(&self, msg: Message) -> Result<()> {
        let partition = self.calculate_partition(&msg);
        let consumers = self.discover_consumers(partition).await?;
        
        let msg = Arc::new(msg);
        
        for consumer in consumers {
            let msg = Arc::clone(&msg);
            tokio::spawn(async move {
                consumer.send((*msg).clone()).await
            });
        }
        
        Ok(())
    }

    fn calculate_partition(&self, msg: &Message) -> u32 {
        // Dynamic partition calculation based on message key
        match &msg.key {
            Some(key) => {
                use std::collections::hash_map::DefaultHasher;
                use std::hash::{Hash, Hasher};
                let mut hasher = DefaultHasher::new();
                key.hash(&mut hasher);
                (hasher.finish() % self.get_partition_count() as u64) as u32
            }
            None => rand::random::<u32>() % self.get_partition_count()
        }
    }

    pub fn get_partition_count(&self) -> u32 {
        // Return a default partition count (you can adjust this value based on your needs)
        32
    }

    async fn discover_consumers(&self, _partition: u32) -> Result<Vec<mpsc::Sender<Message>>> {
        // Implement your consumer discovery logic here
        // - Query registered consumers for the given partition
        // - Return a collection of consumers
        todo!("Implement consumer discovery logic")
    }

    pub async fn start_network(&self) -> Result<()> {
        // Implement network initialization logic here
        // For now, returning Ok to get it compiling
        Ok(())
    }

    pub async fn start_metrics(&self) -> Result<()> {
        // Implement metrics initialization logic here
        // For now, returning Ok to get it compiling
        Ok(())
    }
}
